+++ targeting/tree.go
@@ -195,8 +195,6 @@ func getCondition(c types.TargetingCondition) types.Condition {
 		return conditions.NewTargetExperimentCondition(c)
 	case types.TargetingTargetPersonalization:
 		return conditions.NewTargetPersonalizationCondition(c)
-	case types.TargetingExclusiveFeatureFlag:
-		return conditions.NewExclusiveFeatureFlagCondition(c)
 	case types.TargetingExclusiveExperiment:
 		return conditions.NewExclusiveExperimentCondition(c)
 	case types.TargetingCookie: